[REFACTOR] 디렉토리명 수정에 따른 변경사항
Folder->folder
