+++ src/main/java/com/example/ai_tutor/domain/folder/dto/response/FolderNameListRes.java
@@ -1,4 +1,4 @@
-package com.example.ai_tutor.domain.Folder.dto.response;
+package com.example.ai_tutor.domain.folder.dto.response;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.*;
@@ -8,11 +8,9 @@ import lombok.*;
 @AllArgsConstructor
 @NoArgsConstructor
 @Getter
-public class FolderListRes {
+public class FolderNameListRes {
     @Schema( type = "Long", example = "1", description = "폴더의 id입니다.")
     private Long folderId;
     @Schema( type = "String", example = "빅데이터기술특론", description = "폴더의 이름입니다.")
     private String folderName;
-    @Schema( type = "String", example = "하석재", description = "폴더를 생성한 교수자의 이름입니다.")
-    private String professor;
 }
