+++ src/main/java/com/example/ai_tutor/domain/note/application/StudentNoteService.java
@@ -1,6 +1,6 @@
 package com.example.ai_tutor.domain.note.application;
 
-import com.example.ai_tutor.domain.Folder.domain.repository.FolderRepository;
+import com.example.ai_tutor.domain.folder.domain.repository.FolderRepository;
 import com.example.ai_tutor.domain.note.domain.Note;
 import com.example.ai_tutor.domain.note.domain.repository.NoteRepository;
 
@@ -15,12 +15,6 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-
-
-import java.time.LocalDateTime;
-import java.util.List;
-import java.util.Optional;
-import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
