+++ src/main/java/com/example/ai_tutor/domain/note/domain/repository/NoteRepository.java
@@ -1,6 +1,6 @@
 package com.example.ai_tutor.domain.note.domain.repository;
 
-import com.example.ai_tutor.domain.Folder.domain.Folder;
+import com.example.ai_tutor.domain.folder.domain.Folder;
 import com.example.ai_tutor.domain.note.domain.Note;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
